Checks for corruption earlier and always report errors (#5227)
Was looking into an issue were a mutation was corrupted on the network.
This caused the mutation to be written to the write ahead log and then a
failure occurred in the tablet server which left the tablet in an
inconsistent state.  Modifed the tablet server code to deserialize
mutations as early as possible (it used to deserialize after writing to
the walog, now it does it before).

Wrote an IT to recreate this problem and found another bug.  Writing
data to Accumulo does the following.

 1. Make a startUpdate RPC to create an update session
 2. Make one or more applyUpdates RPCs to add data to the session.  These
    RPCs are thrift oneway calls, so nothing is reported back.
 3. Call closeUpdate on the session to see what happened with all of the
    applyUpdates RPCs done in step 2.

If an unexpected exception happened in step 2 above then it would not be
reported back to the client.  These changes fix and test that as part of
testing the corrupt mutation. After these changes if there was an error
in step 2, then step 3 now throws an exception.