Fixed FileNormalizationIT after #5237
The changes in #5237 to make the external compaction json metadata
entry easier to use ended up breaking FileNormalizationIT. #5237
removed the metadata entry string which ended up escaping the
double quotes in the string for the json and replaced it with an
object that contains the path, start, and end row. The problem
with this is that the metadata entry retains any misnormalization
of the file name (e.g. double slashes in the path) and that path
alone does not. This fixes the issue by putting the metadata entry
into the new object that is serialized to json so that the json
for the external compaction files now contains the metadata entry,
normalized path, start and end row.