Add Mergeability column to support automatic merges (#5187)
Add Mergeability column to support automatic merges

This adds a new Mergeability column for marking if/when tablets are
eligible to me merged by the system based on a threshold. The column 
stores two values, a duration which is a delay for when a tablet can be 
merged that is relative to the time the Manager uses, Steady time. 
It also stores the current Steady time value when inserted so that later 
we can add the delay plus the original time when inserted to see if 
enough time has passed. The Steady Time value will only be stored if 
the delay is >= 0, otherwise it will be null as it won't be used. The Steady
Time value isn't technically needed for a delay of 0 because 0 means
it's eligible to merge always, but it could be useful if we wanted to
change the value for some reason and it would allow logging when it
was stored.

There are 2 possible states for the delay value:

1) null : This means a tablet will never automatically merge
2) duration >= 0 : Tablet is eligible to merge after the given delay
 (stored as a duration), relative to the current system Steady time. Ie. the
 tablet can be merged if the current manager time is later than the delay 
value + the steady time value when inserted. If the duration is 0 then it 
means it can merge always (now). 

This change only adds the new column itself and populates it. The default
is to never merge automatically for all cases except for when the system 
automatically splits tablets. In that case the newly split tablets are marked
as being eligible to merge always (duration of 0).

Future updates will add API enhancements to allow setting/viewing the
mergeability setting as well as to enable automatic merging by the system 
that is based on this new column value.

When automatic merging is enabled, if a user wants to make a tablet 
eligible to be merged in the future they would do so by setting a delay
that is positive. For example, to make a tablet eligible to be merged 
3 days from now the user set a duration of 3 days in the API (future PR)
and when the system inserts the value into metadata it will also include 
the current SteadyTime on creation. Later when the the current steady time 
passes that set delay + original stored steady time value the tablet would be
eligible to be merged. To enable merging immediately they can set the 
duration to 0 or use the TabletMergeability.always() helper which is just a 
shortcut to a duration of 0.