Adds property for jline to prefer using stdout vs stderr (#5221)
* Adds property for jline to always use stdout

When the shell is being used with a `-e <command>` option jline does not
write the output on stdout when also used with a pipe `|` operator.

Instead the output is written to stderr. This causes frustration when
attempting to script actions using the accumulo shell.

* Set system out in TerminalBuilder instead

Change the output behavior manually instead of using the jline
properties.