Propagates errors in Rfile.closeDeepCopies
Rfile.closeLocalityGroupReaders() was suppressing IOExceptions.
This method was called by Rfile.closeDeepCopies() which was called by
FileManager.releaseReaders().  Suppressing the exception meant that
releaseReaders() did not see the exception and would decided to return
the rfile to the pool when it should not.

The only other code calling Rfile.closeLocalityGroupReaders() was
Rfile.close(). Refactored the code so that Rfile.close() still suppressed
the exception and Rfile.closeDeepCopies() does not suppress.  Tried to
preserve the behavior that Rfile.close() closes as many of its
underlying resource as possible even if some exceptions occur.