Use ZooCache detached from Context for ServerPaths
ServiceLockPaths cannot use ClientContext because that can be closed,
and ServiceLockPaths needs to be reused inside the tablet locator thread
pool, which is a static singleton that can survive longer than the
client context.

This removes context from ServiceLockPaths and causes
ZookeeperLockChecker to provide it with a separate root path and
ZooCache.

This fixes the ITs broken after the merge of #5192 in
e745a5dcd4646059f71a026caa277651295f5df0