Offers new ways to compute bulk load plans
Two new ways of computing bulk import load plans are offered in these
changes. First the RFile API was modified to support computing a LoadPlan
as the RFile is written. Second a new LoadPlan.compute() method was
added that creates a LoadPlan from an existing RFile. In addition to
these changes methods were added to LoadPlan that support serializing
and deserializing load plans to/from json.

All of these changes together support the use case of computing load
plans in a distributed manner. For example, with a bulk import directory
with N files the following use case is now supported.

For eack file a task is spun up on a remote server that calls the new
LoadPlan.compute() API to determine what tablets the file overlaps. Then
the new LoadPlan.toJson() method is called to serialize the load plan
and send it to a central place.  All the load plans from the remote
servers are deserialized calling the new LoadPlan.fromJson() method and
merged into a single load plan that is used to do the bulk import.
Another use case these new APIs could support is running this new code
in the map reduce job that generates bulk import data.

In each reducer as it writes to an rfile it could also be building a
LoadPlan. A load plan can be obtained from the Rfile after closing it
and serialized using LoadPlan.toJson() and the result saved to a file.
So after the map reduce job completes each rfile would have
corresponding file with a load plan for that file.  Another process that
runs after the map reduce job can load all the load plans from files and
merge them using the new LoadPlan.fromJson() method. Then the merged
LoadPlan can be used to do the bulk import.  Both of these use cases
avoid doing the analysis of files on a single machine doing the bulk
import. Bulk import V1 had this functionality and would ask random
tservers to do the file analysis. This could cause unexpected load on
those tservers. Bulk V1 would interleave analyzing files and adding them
to tablets. This could lead to odd situations where files are partially
imported to some tablets and analysis fails, leaving the file partially
imported. Bulk v2 does all analysis before any files are added to
tablets, however it lacks this distributed analysis capability. These
changes provide the building blocks to do the distributed analysis that
bulk v1 did for bulk v2.

Co-authored-by: Daniel Roberts <ddanielr@gmail.com>