Support TabletMergeability updates and queries
This is part 2 of the API changes for #5014 and is a follow on to #5187

putSplits() now supports updating existing tablets with the new
TabletMergeability setting, as well as creating new splits.
TabletMergeabilityInfo has been added to TabletInformation so a user can
retrieve the TabletMergeability setting on a tablet as well as the
insertion time and if the tablet is mergeable or not based on the
current manager time.

A couple new RPCs were added to the Manager service including an RPC to
update existing tablets with new TabletMergeability and to retrieve the
current Manager time.