Bumped max watch count from 200 to 300 (#5295)
WatchTheWatchCountIT has been failing in main (again). A recent
change lowered the max watch count down to 200. Running this locally
about 10 times, only one time was it lower than 200. Bumped the max
to 300 to get this test passing. This likely needs bumped to account for
the test running a scan server and/or compactor process.