Enable graceful shutdown of server processes (#5193)
Added RPC to the server processes to initiate a graceful
shutdown. A new admin command has been created
for a user to send this command to a server process.
When the server process receives the command to
shutdown gracefully it will attempt an orderly shutdown.
Additional requests to shutdown gracefully will be
ignored. Compactors will finish the major compaction
that they are currently working on, then will exit. Scan
Servers will return a busy signal to all clients for new
scan requests, and will shutdown when all in-progress
scans have closed. Tablet Servers will signal the Manager
that they are shutting down, which should prevent
assignment of tablets to that server, then they will
unload all tablets and shut down. The Monitor,
Manager, GarbageCollector, and CompactionCoordinator
will shut down also. The last step in the shutdown
process for all servers is to remove their lock in
ZooKeeper.


---------

Co-authored-by: Dom G. <domgarguilo@apache.org>