Refactor Fate lock code to use concrete types
Instead of passing around byte arrays for DistributedReadWriteLock and
related classes we can use LockType and FateId to make the code easier
to work with and to understand as those values are serialized as part of
the lock. A new FateLockEntry type has been created which is passed
around instead and that is serialized/deserialized into the correct byte
array format when reading/writing to zookeeeper.

This closes #5264