adds trace logging of tablet state computations (#5290)
Adds trace level logging of the tablet state and tablet goal state. This
logging will help understand why the manager is making the decisions it
does regarding a particular tablet.  For example if the manager is not
assigning a tablet that one would expect to be assigned turning on this
logging could help.  Because the manager uses an iterator to filter
tablet metadata in the tablet server, may need to look in the log of
manager and tablet servers serving root and metadata tables to see it.

Manually tested this by enabling the trace level logging and runnin some
ITs.  Found a bug with calling TableMetadata.getExtent() in this testing
and corrected that.