Foundation for new Monitor server and UI (#5012)
Created the initial foundation for a new Monitor implementation for
Accumulo 4.0. This change includes a new Thrift RPC endpoint in the
ServerProcess thrift service. The new endpoint returns a MetricResponse
object which contains information about the server and a set of
Micrometer Meters serialized as Google Flatbuffers. The new
Monitor implementation contains a class that periodically calls
this Thrift RPC method on all the server processes. The new
Monitor implementation makes this information, and other specific
information that it gathers available via a new JAX-RS API (see
the Endpoints class).



Co-authored-by: Dom G. <domgarguilo@apache.org>
Co-authored-by: Daniel Roberts ddanielr <ddanielr@gmail.com>
Co-authored-by: Keith Turner <kturner@apache.org>
Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>