Disables HTML escaping in LoadPlan json encoding. (#5299)
The encoding in LoadPlan was using GSon default settings which encoded
some HTML chars like `=` as unicode escape sequences.  This did not
cause any problems because GSon would reverse this encoding.  However it
was not documented in the javadoc for LoadPlan and it seems unnecessary.

Disabled the HTML escaping so the code aligns w/ the javadoc.  Also
updated the unit test to produce `=` in the base64 and see how that
behaves in the json.  This allows the test to better detect changes in
the way the code persist data.