Applied fix from #5260 in other locations (#5305)
The issue fixed in #5260 was that a class that was implementing
a Thrift interface had to declare it, even if it was declared
in a parent class. #5260 fixed the logic in one place and then
removed the interface declarations from the subclasses. It
appears that there are other locations in the code where the
same problem exists. This commit moves the fix to a utility
class and updates the places where this was being done to use
the utility method.