Defers server invalidation in tablet location cache
When many accumulo client threads all decided to invalidate some servers
at the same time in the tablet location cache each threads would scan
though all metadata looking for those servers.

This change adds servers to a set for deferred processing when
invalidated.  This has two advantages.  First for multiple threads
adding servers at around the same time it can lower the number of scans
done over all the cached tablet locations.  Second the amount of work
done to invalidate servers is lowered.  The code used to find cache
entries for a server, add the extent for that cache entry to badExtents,
and then later look for everything in badExtents and remove it.  This
change removes all of that work and directly removes entries from the
cache and avoids adding anything to badExtents.