avoid clearing server in cache when scanner closed and interrupted
When the thread running an accumulo scanner is interrupted it may cause
the server the scanner was reading from to be cleared from the cache.
This can be disruptive in the case where the server is fine. This change
makes a narrow exception to clearing the server from the cache for the
case where the scanner was closed and an interrupt was seen.  The reason
this is so narrow is to avoid failing to invalidate the cache in the
case where there is actually a problem with the server.