Defers server invalidation in tablet location cache (#5308)
Reapplies the changes in in 2.1 from #5308 after ignoring those changes
in the merge commit 9df1a04018f8705a29916bad85de9e0eba969988.  This was
done because the changes were 2.1 were being made to file that was
deleted in main and significantly refactored.  Making these updates in a
stand alone commit makes the diffs easier to look at.