uses consistent zk host addr in MAC (#5316)
Was seeing slow ITs related to recent changes that caused more ZK clients
to be created.  ZK client creation was slow.  The cause was
ZOOKEEPER-3100 and MiniAccumulo setting 127.0.0.1 in ZK server config
and using localhost for ZK in accumulo config. This caused the ZK server
to listen on 127.0.0.1 and the ZK client to sometimes attempt to use the
IPV6 addr for localhost which failed and retried possibly adding seconds
to each ZK connection attempt.

This changes updates accumulo config created by mini accumulo to use
127.0.0.1 for zookeeper.