Filter based on columns declared in TabletMetadataCheck (#5335)
Adds filtering to TabletMetadataCheckIterator that will use matching
families for the provided columns in order to reduce the amount of data
that is fetched. A ResolvedColumns instance is returned by each
TabletMetadataCheck instance that contains the set of columns and set of
matching families to read.

CompactionReservationCheck caches the ResolvedColumns
instance statically to avoid having to recreate the set of columns and
familes for each scan.