Added calls to ZooZap when shutting down groups of servers (#5321)
Modified accumulo-cluster to wait for all spawned ssh commands
to finish and then call ZooZap when shutting down all Compactor,
ScanServer, and TabletServer processes for a resource group.
Modified ZooZap so that the code is consistent for Compactors, 
ScanServers, and TabletServers such that for those server types
it performs a recursive delete at the resource group in the ZK
path.

Closes #5178