Fixed bug in TabletsMetadata.fetch (#5342)
A bug was introduced in #5335 to TabletsMetadata.fetch where the
columns being passed as arguments to the function where not being
used. This has the side effect of fetching all columns in the
TabletMetadata which was causing some ITs to time out. The ITs
in these cases were fetching and counting specific columns in
the TabletMetadata.

This bug also highlighted another bug in CloneIT where an invalid
directory was being inserted into the TabletMetadata which would
then subsequently fail when fetched. The directory column name
was being validated on most, but not all, places where it was being
inserted into the TabletMetadata. I added validation to the
ServerColumnFamily.DIRECTORY_COLUMN.put method and fixed CloneIT.