Started upgrade test, fixed issue in upgradeZookeeper (#5344)
The Upgrader12to13.validateEmptyZKWorkerServerPaths method was
checking for no children in the zookeeper paths for compactors,
scan servers, tablet server, and dead tablet servers as an
indication that no servers were current running. Prior to recent
changes in 4.0, zookeeper paths were not cleaned up for server
processes when those processes are stopped. Modified the logic
in the method to remove any children under those paths that are
empty. Started the Upgrader12to13 unit test.

Closes #5343