Convert several ITs to use SharedMiniClusterBase (#5352)
This change migrates several ITs to using SharedMiniClusterBase so that
a single cluster is stood up and reused for all the tests in the test
class. This has the benefit of speeding up the tests because the entire
cluster does not need to be torn down and recreated for each test.