Fixes metadata balancing issues (#5358)
* Add new metadata balance test to BalanceIT

Adds a new test for checking balancing of the metadata table.
This test currently breaks as there are outstanding bugs in 2.1
related to balancing the metadata table.

Fixes balance related filtering that was using table name instead of id

This change passes a map of tables to balance each balancer.
This reduces the possibility that a balancer will attempt
to balance tables not assigned to itself.

Removes the while loop in favor of just skipping the current datalevel
if specific conditions are met.

---------

Co-authored-by: Keith Turner <kturner@apache.org>
Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>