Modify FateStore to introduce a Seeder API (#5374)
This adds a new Seeder API that can be used to seed transactions with a
FateKey. The purpose of this new API is to allow for eventual batching
of multiple seeding attempts together into one conditional mutation to
improve performance. This change just updates the api and all calls to
the seeder to attempt seeding will be performed individually. A future
update will support the new functionality of batching and commiting all
the changes together when the Seeder is closed.

This change is for Part 1 of #5160