Switches to String-based migration partitioning (#5359)
Move all the partitioned migration code into a separate class to cleanup
the Manager code.

Previously the balancer code would track migrations based on a DataLevel
partition. However, there are cases where a subset of tablets could be
handled by a sub-balancer. This change supports separating the tablets
by a String value which allows for more flexibility.

Changes the MetaDataTableScanner to pull migrationSnapshot data based on
the datalevel of the target table.

Use the .Name() method for the DataLevel vs a hardcoded string value

Removed the RootTabletStateStore as the addition of DataLevels has
removed the need for specifying a different table name for the
MetaDataTableScanner.

---------

Co-authored-by: Keith Turner <kturner@apache.org>
Co-authored-by: Christopher L. Shannon <cshannon@apache.org>