prevents creating overlapping encryption streams
Modified AESCryptoService to detect creation of multiple encryptions
streams at the same time. The implementation in AESCryptoService does
not support this because it uses shared Cipher object and shrared
intialization vectors.  The code that uses CryptoService does not seem
to have a need for creating more that one encryption stream at a time.
If the usage pattern of the code change, then the detection added in
this commit should catch that.

fixes #5386