Fixed UpgradeCoordinator, UpgradeProgressTracker, and Upgrader12to13
Updated the UpgradeCoordinator to fix what is essentially an off by
one bug. The version of the instance and the key in the upgraders
map did not match the way the code was processing the versions.

Updated the UpgradeProgressTracker with additional checks.

Updated Upgrader12to13 methods to be invoked during the appropriate
phase and to be re-entrant.