Enable auto merging of tablets (#5353)
This change adds support for periodically scanning tables to find ranges
of tables that can be automatically merged. The thread that runs uses
the recently added TabletMergeabilty column as well as total file
counts/sizes to compute if a range can be merged. By default, tablets
can be merged if the total size is less than 25% of the split threshold
so we don't immediately merge only to split again.

There is a new thread in the Manager class that performs the computation
and will submit fate jobs for merging on tablet ranges. There is a new
fate operation during merge that will validate the range is still ok to
merge if the merge was submitted as a system merge.

This closes #5014