Create custom TMuliplexedProcessor to block Thrift calls when needed
Copied TMultiplexedProcessor to a new class, HighlyAvailableServiceTMultiplexedProcessor,
and added the logic from HighlyAvailableServiceInvocationHandler to it so
that an error is returned when the service is not active or upgrading.

This moves the location where the exception is thrown up the call stack and
out of ProcessFunction where the incorrect logging is occurring.

Closes #5411