removes root path from lock id serialization (#5417)
In order to make this change needed to analyze how the code uses LockId,
based on this anylsis ended up making a few changes.

Some code was caching the serialized form of the lock id and keeping a
string around.  Its better to keep the lock id type around instead of a
string, so made lock id memoize serialization to support doing this.

The lock id constructor was removing extraneous slashes at the end of
the lock path.  However it seemed this would never happen in the code.
Removed this handling and instead added a check that paths do not end w/
a slash.  Also added some other validation in the constructor. Noticed
this object is created alot in the manager and tablet server, to avoid
the repeated validation made ServiceLock cache the LockId instead of
always creating it.

fixes #5410


Co-authored-by: Christopher Tubbs <ctubbsii@apache.org>