Don't start Thrift servers until upgrade done and lock acquired
Removed HighlyAvailableService and associated code, which was only
really used by the Manager, and instead modified the Manager to
only advertise the Manager address with the Thrift services after
the upgrade process has completed and the Manager is fully started.

This change removes the HighlyAvailableServiceInvocationHandler
which was throwing an exception when the Manager was not ready
to accept Thrift RPC calls and causing stack traces in the Manager
log during the upgrade process. These stack traces could be
construed as a problem to the user, when in reality they were
benign.

Closes #5411