Don't advertise Manager address until upgrade done and fully started (#5419)
Removed HighlyAvailableService and associated code, which was only
really used by the Manager, and instead modified the Manager to
only advertise the Manager address with the Thrift services after
the upgrade process has completed and the Manager is fully started.

This change removes the HighlyAvailableServiceInvocationHandler
which was throwing an exception when the Manager was not ready
to accept Thrift RPC calls and causing stack traces in the Manager
log during the upgrade process. These stack traces could be
construed as a problem to the user, when in reality they were
benign.

Also modified TServerUtils to only create the Thrift server, but not
start it. The ThriftServer is set on the ServerAddress object that is
return by the TServerUtils methods. Added a method to ServerAddress
to start the Thrift server, which for most server processes is called
immediately after creating it. However, for the Manager, the call
to start the Thrift server is delayed until the Manager is up.

Closes #5411