Process seeding of split fate operations in batches (#5404)
Updates the Seeder in the Manager that handles seeding split fate ops to
use a single thread and to submit multiple outstanding operations to
be seeded together instead of individually in order to improve
performance. The user fate store will now track outstanding fate
operations and return a future for each pending operation that will be
completed when the batch is submitted.

This closes #5160