Fixes performance bug in GroupBalancer (#5432)
Code in GroupBalancer was calling
`HashBasedTable.columnKeySet().removeAll(aList)`.  The Guava
HashBasedTable code would iterator over all of its rows and columns and
then call contains on the list passed in by GroupBalancer. So this code
was doing `ROWS*COLS*LIST_SIZE` operations.   Changed the list to a
hashset in this PR so now it only does `ROW*COLS` operations.