Recreate TabletsMetadata iterator when file ranges are not contiguous (#5341)
In the Bulk Import v2 LoadFiles step a single TabletsMetadata
object was used to map a tables tablets to a set of bulk import
files. In the case where a small percentage of tablets were
involved in the bulk import a majority of the tables tablets
would still be evaluated. In the case where bulk imports were
not importing into contiguous tablets the code would just
iterate over the tables tablets until it found the next starting
point.

This change creates a new  property to determine how many
TabletMetadata objects to evaluate before recreating the
TabletsMetadata object at the new start position. When bulk
importing into a set of contiguous tablets or into a majority
of tablets in a table, the value of this property should likely
be zero to leave this feature disabled. When importing into
a non-contiguous set of tablets, then this property could be
used to speed up the bulk import processing.

This change also closes the TabletsMetadata objects which was not
being done previously.

Related to #5201



Co-authored-by: Daniel Roberts <ddanielr@gmail.com>